+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 
+### New Features
+- **Update Notifications**: A system notification is now sent on startup if a new version of Proximity is available.
+
 ### Breaking Changes
 - **Profiles**: Added support for using multiple use-cases. Profiles define an Atlassian Cloud ID, use case ID, and AD group. Each let's you route requests to AI-Gateway using it's own configuration.